+++ R/bind_each_to_all.R
@@ -1,11 +1,11 @@
-#' Bind each to all
+#' Bind each all
 #'
 #' @description
 #' Bind data to support comparing each category to all of the data.
 #'
 #' @param data A data frame or tibble.
 #' @param ... An unquoted variable.
-#' @param name A variable name. Defaults to `each_or_all`.
+#' @param name A variable name. Defaults to `each_all`.
 #' @param each A string for the each value. Defaults to `"Each"`.
 #' @param all A string for the all value. Defaults to `"All"`.
 #' @param all_after A number for where the all value should be placed after. Use 0 for first or Inf for last. Defaults to Inf.
@@ -21,34 +21,34 @@
 #' set_blanket()
 #'
 #' penguins |>
-#'   bind_each_to_all(species) |>
-#'   distinct(species, each_or_all)
+#'   bind_each_all(species) |>
+#'   distinct(species, each_all)
 #'
 #' penguins |>
-#'   bind_each_to_all(species) |>
+#'   bind_each_all(species) |>
 #'   gg_jitter(
 #'     x = species,
 #'     y = body_mass_g,
 #'   )
 #'
 #' penguins |>
-#'   bind_each_to_all(species) |>
+#'   bind_each_all(species) |>
 #'   gg_jitter(
 #'     x = species,
 #'     y = body_mass_g,
-#'     col = each_or_all,
+#'     col = each_all,
 #'     col_palette = c(blue, grey),
 #'   ) +
 #'   theme(legend.position = "none")
 #'
 #' penguins |>
-#'   bind_each_to_all(species) |>
-#'   group_by(species, each_or_all) |>
+#'   bind_each_all(species) |>
+#'   group_by(species, each_all) |>
 #'   summarise(across(body_mass_g, \(x) mean(x, na.rm = TRUE))) |>
 #'   gg_col(
 #'     x = species,
 #'     y = body_mass_g,
-#'     col = each_or_all,
+#'     col = each_all,
 #'     col_palette = c(blue, grey),
 #'     width = 0.5,
 #'     y_label = "Average body mass g",
@@ -56,13 +56,13 @@
 #'   theme(legend.position = "none")
 #'
 #' penguins |>
-#'   bind_each_to_all(species, all = "All\nspecies") |>
+#'   bind_each_all(species, all = "All\nspecies") |>
 #'   gg_jitter(
 #'     x = species,
 #'     y = body_mass_g,
-#'     col = each_or_all,
+#'     col = each_all,
 #'     col_palette = c(blue, grey),
-#'     facet = each_or_all,
+#'     facet = each_all,
 #'     facet_layout = "grid",
 #'     facet_scales = "free_x",
 #'     facet_space = "free_x",
@@ -71,9 +71,9 @@
 #'   theme(strip.text.x = element_blank()) +
 #'   labs(x = NULL)
 #'
-bind_each_to_all <- function(data,
+bind_each_all <- function(data,
                         ...,
-                        name = "each_or_all",
+                        name = "each_all",
                         each = "Each",
                         all = "All",
                         all_after = Inf) {